put testbanker in its own file
