+++ tests/file.go
@@ -12,7 +12,6 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/gnolang/gno/pkgs/crypto"
 	gno "github.com/gnolang/gno/pkgs/gnolang"
 	osm "github.com/gnolang/gno/pkgs/os"
 	"github.com/gnolang/gno/pkgs/std"
@@ -438,68 +437,4 @@ func trimTrailingSpaces(result string) string {
 		lines[i] = strings.TrimRight(line, " \t")
 	}
 	return strings.Join(lines, "\n")
-}
-
-// ----------------------------------------
-// testBanker
-
-type testBanker struct {
-	coinTable map[crypto.Bech32Address]std.Coins
-}
-
-func newTestBanker(args ...interface{}) *testBanker {
-	coinTable := make(map[crypto.Bech32Address]std.Coins)
-	if len(args)%2 != 0 {
-		panic("newTestBanker requires even number of arguments; addr followed by coins")
-	}
-	for i := 0; i < len(args); i += 2 {
-		addr := args[i].(crypto.Bech32Address)
-		amount := args[i+1].(std.Coins)
-		coinTable[addr] = amount
-	}
-	return &testBanker{
-		coinTable: coinTable,
-	}
-}
-
-func (tb *testBanker) GetCoins(addr crypto.Bech32Address) (dst std.Coins) {
-	return tb.coinTable[addr]
-}
-
-func (tb *testBanker) SendCoins(from, to crypto.Bech32Address, amt std.Coins) {
-	fcoins, fexists := tb.coinTable[from]
-	if !fexists {
-		panic(fmt.Sprintf(
-			"source address %s does not exist",
-			from.String()))
-	}
-	if !fcoins.IsAllGTE(amt) {
-		panic(fmt.Sprintf(
-			"source address %s has %s; cannot send %s",
-			from.String(), fcoins, amt))
-	}
-	// First, subtract from 'from'.
-	frest := fcoins.Sub(amt)
-	tb.coinTable[from] = frest
-	// Second, add to 'to'.
-	// NOTE: even works when from==to, due to 2-step isolation.
-	tcoins, _ := tb.coinTable[to]
-	tsum := tcoins.Add(amt)
-	tb.coinTable[to] = tsum
-}
-
-func (tb *testBanker) TotalCoin(denom string) int64 {
-	panic("not yet implemented")
-}
-
-func (tb *testBanker) IssueCoin(addr crypto.Bech32Address, denom string, amt int64) {
-	coins, _ := tb.coinTable[addr]
-	sum := coins.Add(std.Coins{{denom, amt}})
-	tb.coinTable[addr] = sum
-}
-
-func (tb *testBanker) RemoveCoin(addr crypto.Bech32Address, denom string, amt int64) {
-	coins, _ := tb.coinTable[addr]
-	rest := coins.Sub(std.Coins{{denom, amt}})
-	tb.coinTable[addr] = rest
 }
